requirements.txt
aiofiles==24.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Cython==3.0.11
fastapi==0.115.6
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.19.1
numpy==1.26.4
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.0.0
pmdarima==2.0.4
protobuf==5.29.2
pyarrow==18.1.0
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.3
statsmodels==0.14.4
streamlit==1.41.1
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
